Updated TaskProvider mechanism for finding simulation names
The TaskProvider needs to find all simulation names in an .apsimx file.
Previously, we did this by parsing the xml in the .apsimx file but this
will not be feasible now since we have moved to json for the file
format. Moreover, the fileformat (gradually) changes over time, which
could break any parsing code we do write. Moreover again, simulation
names cannot always be deduced from the .apsimx file itself, as in the
case of an Experiment, or, in a more complex situation, a sensitivity
analysis model which will need to run R code in order to determine
which simulations to run.

This is a bit of a hacky solution wherein we assume that this work has
already been done, and an .apsimx file exists for each simulation to be
run (equivalent to "Generate .apsimx files" in the GUI). For each
.apsimx file we find, just return the name of the file and this is
assumed to be the same as the simulation name. For each simulation
found, we generate an azure task to be run on an execution node which
will basically call Models.exe to run the given .apsimx file. This is
problematic in a number of ways but for now I just want to get this
azure stuff actually working. In the future, maybe we can merge this
azure job manager into the apsimx source code, or call Models.exe (or
another apsim executable) to query the .apsimx file for a list of
simulation names to run.